Initial creation.
This project will contain my python filter code. It has previously
lived in various places, mostly in my Kalman filtering book repository,
but it deserves to be it's own project.

This checks in the LeastSquares filter code, but that is secondary
to the idea of just creating the project and getting it up on
Github. More files to follow.