Added batch processing, and RKS smoother.
Added batch processing to KalmanFilter class to filter an entire set
of measurements at once.

Implemented a Rauch-Tung-Striebel (RKS) smoother. I am not at all sure
it is correct yet. Output is mostly the same as the Kalman filter, just
slightly smoother. However, KF must be initialized with a good value for
P, or the smoothing step goes unstable. Not sure if that is a computation
issue due to singularity, or what. a lot more testing to be done before
I am happy with this class.

Also, it probably shouldn't be a class, but a function.