Added rks algorithm back.
This is in a weird state. It was created to do RKS smoothing. For some
reason I don't recall, I ended up putting a fixed lag smoother in there
in front of the RKS code, so RKS was not being done. Now RKS is back,
but the fixed lag is still there - needs to be moved out.