Removed or disabled incomplete code
The information filter was copied from the Kalman filter code.
Many of the helper functions are not implemented - I have put
exceptions at the top of the function to ensure that they are
not used.

There was also an Extended Kalman filter which I just deleted.
I'm sure I need to replace it with an extended Information filter,
but I am not prepared to do that yet.