Fixed import error.
I do not understand this, but if importing unscented_transform.py is
the last line then when you try to import the function it is treated
like a module and you get an error. Putting it in the middle of the
import lets it work correctly. I have no idea if this is a python bug
or a limitation in my understanding.