Removed checking for zs in batch_filter.
the check was not particularly good. It assumed that the first
element in zs is representative of all elements in the list.
However, Python lists are heterogeneous. And this is encouraged
by the API - you can optionally use `None` for any value.

The upshot is that if the first element of zs was None the program
raised an exception.

There is no real utility to this check - any element of zs could
be bad, and that is checked in the call to update() anyway.