Changed log-likelihood to a method
Computation of log-liklihood on every update is very expensive;
it was cutting the speed of the Kalman filter in half. Now you
must call a method which computes it for you. This also has the
advantage of being able to compute the likelihood of any measurement,
not just the one you pass to update()