fixed float64.dot() call
The functions update and predict allow you to use floats instead
of matrices to implement the univariate case. However, there were
forms of dot(a,b).dot(c) in the code, which raise exceptions if
all the variables are floats. Changed to the nested form
dot(dot(a,b),c)