GitHub #138
I wasn't saving any attributes when update() was called with Zone.

Altered self.z to be a column vector of None with (dim_z, 1) shape
so that code using the Saver class can be more consistent.

Added saving of posterior and z to all classes to mirror how
KalmanFilter does it.