Compute log-likehood only if accessed #139
Previously I modified KalmanFilter to use None to denote that
the log-likeliood needed to be computed, and added properties
to compute on demand. This is because the computation of the logpdf
is quite expensive.

This check in normalizes that behavior across all of the classes.

I also added mahalanobis distance to the classes which didn't
implement it.