Multiple fixes
Unfortunate monolithic commit.

* Partially removed dependence on Matplotlib. Matplotlib is only imported
by functions that use plotting. Github #148

* Removed ability to pass in single matrix to batch_filter functions/methods.
Github #147. In general it is impossible to tell if you have a list of matrices,
or an single array, depending on the shape of the arrays and the size of the
input. Rather than try and obscurely fail, make the user pass the correctly
dimensioned data in.

* various fixes typo cleanups